Исправил согласно замечаниям.
