+++ GitHubApp.xcworkspace/xcuserdata/colonelcassad.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -288,7 +288,7 @@
             endingColumnNumber = "9223372036854775807"
             startingLineNumber = "16"
             endingLineNumber = "16"
-            landmarkName = "didUserHasToken()"
+            landmarkName = "userHasToken()"
             landmarkType = "7">
          </BreakpointContent>
       </BreakpointProxy>
