+++ GitHubApp/ViewControllers/Extensions/SearchReposViewController+FaceIdChecking.swift
@@ -20,7 +20,7 @@ extension SearchReposViewController {
         helloLabel.text = "Hello, \(currentUser.login)!"
         avatarImage.kf.setImage(with: URL(string: currentUser.avatarURL))
       } else {
-        networkObject.getCurrentUser(token: token) {[weak self] user in
+        networkObject.getCurrentUser(token: token) { [weak self] user in
           guard let self = self, let user = user else {return}
           DispatchQueue.main.async {
             self.user = user
@@ -30,7 +30,7 @@ extension SearchReposViewController {
       
     } else {
       if let strigyfiedCode = String(data: tempCode, encoding: .utf8) {
-        networkObject.getAuthorizationToken(code: strigyfiedCode) {[weak self] data in
+        networkObject.getAuthorizationToken(code: strigyfiedCode) { [weak self] data in
           guard let self = self else {return}
           guard let data = data else {return}
           guard let token = Parser.getToken(data) else
@@ -41,7 +41,7 @@ extension SearchReposViewController {
             return
           }
           let _ = KeyChainService.save(key: "accessToken", data: token)
-          networkObject.getCurrentUser(token: token) {user in
+          networkObject.getCurrentUser(token: token) { user in
             DispatchQueue.main.async {
               self.user = user
             }
@@ -52,8 +52,8 @@ extension SearchReposViewController {
   }
   
   func performFaceIdCheck()  {
-    let b_securityChecker = TFID()
-    b_securityChecker.authenticateUser{[weak self] isRecognized in
+    let biometrySecurityChecker = FaceIdWrapper()
+    biometrySecurityChecker.authenticateUser { [weak self] isRecognized in
       guard let self = self else
       {return}
       if isRecognized{
